added automatic_scale_choice, analytic_check default value, fixed tf float precision

+++ n3fit/src/n3fit/layers/CombineCfac.py
@@ -123,7 +123,7 @@ class CombineCfacLayer(Layer):
                 dtype=np.float32,
             )
             linear = self._compute_linear(linear_values)
-            inputs = tf.cast(inputs, tf.float64)
+            inputs = tf.cast(inputs, tf.float32)
 
             return (1 + linear) * inputs
